requirements.txt
pandas==1.5.3
numpy==1.23.5
matplotlib==3.7.2
seaborn==0.12.2
streamlit==1.24.0
scikit-surprise==1.1.3
joblib==1.3.2